# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: SystemNavigator AI API
  version: 0.3.0
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/dead-letters:
    get: {summary: List tenant dead letters for administrators, responses: {"200": {description: OK}, "403": {description: Forbidden}}}
  /api/v1/admin/dead-letters/{job_id}/retry:
    post: {summary: Requeue dead letter as administrator, responses: {"200": {description: Queued}, "403": {description: Forbidden}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT